stcuk as hell

+++ http/controllers/task.controller.ts
@@ -7,6 +7,11 @@ import { JSON_MESSAGES } from "./utils";
 class TaskController {
   async fetchUserTasksList(req: Request, res: Response) {
     const { userId } = req.params;
+    if (!userId) {
+      return res
+        .status(400)
+        .json({ success: false, message: JSON_MESSAGES.BAD_REQUEST });
+    }
     const limit = parseInt(req.body.limit) || 10;
     const page = parseInt(req.body.page) || 1;
     const tasks = await TasksServiceInstance.getList(limit, page, userId);
@@ -15,6 +20,11 @@ class TaskController {
 
   async addUserTask(req: Request, res: Response) {
     const { userId } = req.params;
+    if (!userId) {
+      return res
+        .status(400)
+        .json({ success: false, message: JSON_MESSAGES.BAD_REQUEST });
+    }
     const task = await TasksServiceInstance.create(req.body, userId);
     appDevelopmentLogger({ task }, { context: "Task Create Controller" });
     if (!task) {
@@ -27,6 +37,11 @@ class TaskController {
 
   async updateUserTaskById(req: Request, res: Response) {
     const { taskId, userId } = req.params;
+    if (!userId || !taskId) {
+      return res
+        .status(400)
+        .json({ success: false, message: JSON_MESSAGES.BAD_REQUEST });
+    }
     const task = await TasksServiceInstance.updateById(
       taskId,
       req.body,
@@ -42,6 +57,11 @@ class TaskController {
 
   async fetchUserTask(req: Request, res: Response) {
     const { taskId, userId } = req.params;
+    if (!userId || !taskId) {
+      return res
+        .status(400)
+        .json({ success: false, message: JSON_MESSAGES.BAD_REQUEST });
+    }
     const task = await TasksServiceInstance.getById(taskId, userId);
     if (!task) {
       return res
@@ -53,6 +73,11 @@ class TaskController {
 
   async deleteUserTask(req: Request, res: Response) {
     const { taskId, userId } = req.params;
+    if (!userId || !taskId) {
+      return res
+        .status(400)
+        .json({ success: false, message: JSON_MESSAGES.BAD_REQUEST });
+    }
     const isDeleted = await TasksServiceInstance.deleteById(taskId, userId);
     if (!isDeleted) {
       return res
